Implemented Basic Triggers
The module now returns a basic message when a form is submitted.
Additionally, I attempted to assign the array of results from the form
to another array but it is only returning a null array.  More work is
required for that functionality -- it is possible that
entityform_get_submissions is the incorrect function, or is firing at an
incorrect time.