Module Functions at Basic Level
The module correctly retrieves values from the form entries, then
converts those values into drush commands (set by convention -- name of
question corresponds to drush command). The drush commands are written
to a file (which is named after the form title) and saved in the Drupal7
directory (as a .drecipe file). Several remaining QOL fixes are needed
to get the module to perfectly handle T/F forms and properly convert
them. The To-Be-Done comment section lists the remaining tasks before
the module is expanded to handle different types of questions.